+++ packages/core/src/graph/graph.ts
@@ -37,7 +37,7 @@ import { lerp, move } from '../util/math';
 import { dataValidation, singleDataValidation } from '../util/validation';
 import Global from '../global';
 import { ItemController, ModeController, StateController, ViewController } from './controller';
-import { plainCombosToTrees, traverseTree, reconstructTree, traverseTreeUp } from '../util/graphic';
+import { plainCombosToTrees, traverseTree, reconstructTree, traverseTreeUp, getAnimateCfgWithCallback } from '../util/graphic';
 import Hull from '../item/hull';
 
 const { transform } = ext;
@@ -224,7 +224,7 @@ export default abstract class AbstractGraph extends EventEmitter implements IAbs
       /**
        * Minimum scale size
        */
-      minZoom: 0.2,
+      minZoom: 0.02,
       /**
        * Maxmum scale size
        */
@@ -553,36 +553,12 @@ export default abstract class AbstractGraph extends EventEmitter implements IAbs
    * @param {string} state 状态
    * @return {object} 元素实例
    */
-  public findAllByState<T extends Item>(type: ITEM_TYPE, state: string): T[] {
-    return this.findAll(type, item => item.hasState(state));
-  }
-
-  private getAnimateCfgWithCallback({
-    animateCfg,
-    callback
-  }: {
-    animateCfg: GraphAnimateConfig;
-    callback: () => void;
-  }): GraphAnimateConfig {
-    let animateConfig: GraphAnimateConfig;
-    if (!animateCfg) {
-      animateConfig = {
-        duration: 500,
-        callback
-      };
+  public findAllByState<T extends Item>(type: ITEM_TYPE, state: string, additionalFilter?: (item: Item) => boolean): T[] {
+    if (additionalFilter) {
+      return this.findAll(type, item => item.hasState(state) && additionalFilter(item));
     } else {
-      animateConfig = clone(animateCfg);
-      if (animateCfg.callback) {
-        const animateCfgCallback = animateCfg.callback;
-        animateConfig.callback = () => {
-          callback();
-          animateCfgCallback();
-        }
-      } else {
-        animateConfig.callback = callback;
-      }
+      return this.findAll(type, item => item.hasState(state));
     }
-    return animateConfig;
   }
 
   /**
@@ -600,7 +576,7 @@ export default abstract class AbstractGraph extends EventEmitter implements IAbs
       matrix = [1, 0, 0, 0, 1, 0, 0, 0, 1];
     }
     if (animate) {
-      const animateConfig = this.getAnimateCfgWithCallback({
+      const animateConfig = getAnimateCfgWithCallback({
         animateCfg,
         callback: () => this.emit('viewportchange', { action: 'translate', matrix: group.getMatrix() })
       });
@@ -728,28 +704,31 @@ export default abstract class AbstractGraph extends EventEmitter implements IAbs
    */
   public zoom(ratio: number, center?: Point, animate?: boolean, animateCfg?: GraphAnimateConfig): boolean {
     const group: IGroup = this.get('group');
-    let matrix = clone(group.getMatrix());
+    let matrix = clone(group.getMatrix()) || [1, 0, 0, 0, 1, 0, 0, 0, 1];
     const minZoom: number = this.get('minZoom');
     const maxZoom: number = this.get('maxZoom');
+    const currentZoom = this.getZoom() || 1;
+    const targetZoom = currentZoom * ratio;
+    let finalRatio = ratio;
 
-    if (!matrix) {
-      matrix = [1, 0, 0, 0, 1, 0, 0, 0, 1];
+    let failed = false;
+    if (minZoom && targetZoom < minZoom) {
+      finalRatio = minZoom / currentZoom;
+      failed = true;
+    } else if (maxZoom && targetZoom > maxZoom) {
+      finalRatio = maxZoom / currentZoom;
+      failed = true;
     }
 
     if (center) {
       matrix = transform(matrix, [
         ['t', -center.x, -center.y],
-        ['s', ratio, ratio],
+        ['s', finalRatio, finalRatio],
         ['t', center.x, center.y],
       ]);
     } else {
-      matrix = transform(matrix, [['s', ratio, ratio]]);
+      matrix = transform(matrix, [['s', finalRatio, finalRatio]]);
     }
-
-    if ((minZoom && matrix[0] < minZoom) || (maxZoom && matrix[0] > maxZoom)) {
-      return false;
-    }
-    // matrix = [2, 0, 0, 0, 2, 0, -125, -125, 1];
 
     if (animate) {
       // Clone the original matrix to perform the animation
@@ -758,9 +737,9 @@ export default abstract class AbstractGraph extends EventEmitter implements IAbs
         aniMatrix = [1, 0, 0, 0, 1, 0, 0, 0, 1];
       }
       const initialRatio = aniMatrix[0];
-      const targetRatio = initialRatio * ratio;
+      const targetRatio = initialRatio * finalRatio;
 
-      const animateConfig = this.getAnimateCfgWithCallback({
+      const animateConfig = getAnimateCfgWithCallback({
         animateCfg,
         callback: () => this.emit('viewportchange', { action: 'zoom', matrix: group.getMatrix() })
       });
@@ -785,7 +764,7 @@ export default abstract class AbstractGraph extends EventEmitter implements IAbs
       this.autoPaint();
     }
 
-    return true;
+    return !failed;
   }
 
   /**
@@ -2459,7 +2438,7 @@ export default abstract class AbstractGraph extends EventEmitter implements IAbs
    * 若 cfg 含有 type 字段或为 String 类型，且与现有布局方法不同，则更换布局
    * 若 cfg 不包括 type ，则保持原有布局方法，仅更新布局配置项
    */
-  public updateLayout(cfg: any, align?: 'center' | 'begin', alignPoint?: IPoint): void {
+  public updateLayout(cfg: any, align?: 'center' | 'begin', alignPoint?: IPoint, stack: boolean = true): void {
     const layoutController = this.get('layoutController');
 
     if (isString(cfg)) {
@@ -2501,7 +2480,7 @@ export default abstract class AbstractGraph extends EventEmitter implements IAbs
       }
     }
 
-    const oriLayoutCfg = this.get('layout');
+    const oriLayoutCfg = { ...this.get('layout') };
     const layoutCfg: any = {};
     Object.assign(layoutCfg, oriLayoutCfg, cfg);
     this.set('layout', layoutCfg);
@@ -2517,6 +2496,10 @@ export default abstract class AbstractGraph extends EventEmitter implements IAbs
     } else {
       // has different type, change layout
       layoutController.changeLayout(layoutCfg);
+    }
+
+    if (stack && this.get('enabledStack')) {
+      this.pushStack('layout', { before: oriLayoutCfg, after: layoutCfg });
     }
   }